restore: implement DeltaRestoreOperations methods
Longhorn 6576
